fix: catching results with google search errors for resuming
When analyzing the results, I find that some records are actually failed due to google search errors, but they are not explicitly labelled as errors because there is still a dummy result like "I did not find any results due to google search errors". This new commit catches these errors and rerun them, when evlauation scripts are called with `--resume-from` parameters.